requirements.txt
ansi2html==1.9.2
azure_storage==0.37.0
boto3==1.42.32
colorama==0.4.6
deepdiff==9.0.0
jsonlines==4.0.0
matplotlib==3.10.8
numpy==2.4.4
openai==2.30.0
pandas==3.0.2
protobuf==7.34.1
python_dateutil==2.9.0.post0
rich==14.3.3
scikit_learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
